[pages] all added
