--- vite-messenger/src/components/Button/button.css
@@ -1,6 +1,0 @@
-.btn{
-    border-radius: 20px;
-    padding: 20px 40px;
-    color: forestgreen;
-    background: black;
-}
