+++ vite-messenger/src/components/Form/index.ts
@@ -1,29 +1,9 @@
-import Block from '../../utils/Block';
-import styles from './button.css'
 
-interface ButtonProps {
-  label: string;
-  type?: 'submit' | 'button',
-  onClick?: () => void;
-  events: {
-    click: () => void;
-  };
-}
+import template from "./form.hbs";
+import Block from "../../utils/Block";
 
-export class Form extends Block<ButtonProps> {
-  constructor(props: ButtonProps) {
-    super({
-      ...props,
-      events: {
-        click: props.onClick
-      }
-    });
-    // НАВЕШИВАЕМ КЛАСС НА КНОПКУ
-    // this.element!.classList.add(styles.btn)
-  }
-
-
-  render() {
-    return this.compile(`{{label}}`, this.props);
-  }
+export default class Form extends Block {
+    render() {
+        return this.compile(template, this.props);
+    }
 }
