+++ vite-messenger/src/components/Profile/changeItem.hbs
@@ -1,0 +1,4 @@
+<div class="profile-data__row" id="profile-data">
+    <label class="bold">{{first}}</label>
+    <input name="{{name}}" class="second"  value="{{second}}" />
+</div>
