+++ vite-messenger/src/components/Profile/profileMain.hbs
@@ -31,12 +31,17 @@
 
     <div class="profile-main__buttons">
 
-        <button>
-            <a href="/change">
-                Изменить данные
-            </a>
-
-        </button>
+<!--        <button>-->
+<!--            <a href="/change">-->
+<!--                Изменить данные-->
+<!--            </a>-->
+<!--                -->
+<!--        </button>-->
+        {{{Button
+                onClick=this.onClickChange
+                label="Изменить данные"
+                type="button"
+        }}}
         <hr>
         <button>Изменить пароль</button>
         <hr>
