+++ vite-messenger/src/pages/error/error.hbs
@@ -1,0 +1,5 @@
+<main>
+    <h1>500</h1>
+    <p>Мы уже фиксим</p>
+    <a href="/">Назад </a>
+</main>
