+++ vite-messenger/src/pages/profile/profileStyles.pcss
@@ -8,7 +8,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    img{
+    a img{
         width: 28px;
         height: 28px;
     }
