--- vite-messenger/src/testComponents/Form/index.ts
@@ -1,9 +1,0 @@
-
-import template from "./form.hbs";
-import Block from "../../utils/Block";
-
-export default class Form extends Block {
-    render() {
-        return this.compile(template, this.props);
-    }
-}
