+++ vite-messenger/src/utils/render.ts
@@ -4,12 +4,16 @@ import {AuthPage} from "../pages/auth";
 import {RegistrationPage} from "../pages/registration";
 import {ChatPage} from "../pages/chat";
 import {ProfilePage} from "../pages/profile";
+import {ChangeData} from "../pages/changeData";
+import {ErrorPage} from "../pages/error";
 
 const ROUTES = {
     'auth': AuthPage,
     'reg': RegistrationPage,
     'chat': ChatPage,
     'profile': ProfilePage,
+    'change': ChangeData,
+    'error': ErrorPage,
 }
 
 export function render(name: keyof typeof ROUTES) {
